requirements.txt
asn1crypto==0.24.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
dnspython==1.16.0
docopt==0.6.2
eventlet==0.25.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.3.2
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-bigquery==1.20.0
google-cloud-core==1.0.3
google-cloud-datastore==1.7.1
google-cloud-pubsub==1.0.2
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
greenlet==0.4.15
grpc-google-iam-v1==0.12.3
grpcio==1.24.0
gunicorn==19.9.0
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
oauth2client==4.1.2
parso==0.5.1
prompt-toolkit==2.0.10
protobuf==3.9.2
psycopg2==2.8.3
ptpython==2.0.4
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
Pygments==2.4.2
pymongo==3.7.2
PyMySQL==0.9.2
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.11.0
SQLAlchemy==1.3.8
urllib3==1.25.6
uWSGI==2.0.18
wcwidth==0.1.7
Werkzeug==0.16.0
